fix(phase5): correct P0 sync word to 0x12 (private, 1-byte convention)
ral_set_lora_sync_word() takes a uint8_t in the SX127x/SX126x 1-byte
convention: 0x12 = private network, 0x34 = public network.  The SX126x
chip driver (sx126x_set_lora_sync_word) computes the 2-byte register
encoding (0x12 -> 0x1424) internally.

The table previously stored 0x1424 (the 2-byte register value), which
when narrowed to uint8_t gave 0x24 -- a valid but wrong sync word that
would not match GRC's loramac-node (which uses 0x12 = private, i.e.
public_network=false in LoRa.cpp).

Fix: store 0x12 in all LoRa profiles; add explicit uint8_t cast in
applyLoRa; update comment + gtest to reflect the 1-byte convention.

+++ fprime-zephyr/Drv/UspRadio/RalSessionImpl.cpp
@@ -218,8 +218,9 @@ int RalSessionImpl::applyLoRa(const LoRaParams& p) {
     pkt.invert_iq_is_on      = false;
     if (ral_set_lora_pkt_params(m_ral, &pkt) != RAL_STATUS_OK) return -EIO;
 
-    // 4. Sync word
-    if (ral_set_lora_sync_word(m_ral, p.sync_word) != RAL_STATUS_OK) return -EIO;
+    // 4. Sync word (1-byte convention: 0x12 = private, 0x34 = public;
+    //    ral_set_lora_sync_word takes uint8_t; LinkProfiles stores the value in U16)
+    if (ral_set_lora_sync_word(m_ral, static_cast<uint8_t>(p.sync_word)) != RAL_STATUS_OK) return -EIO;
 
     // 5. Frequency + PA config
     if (ral_set_rf_freq(m_ral, m_freq_hz) != RAL_STATUS_OK) return -EIO;
